src/python/ros_adapter/requirements.txt
google-auth==1.4.1
google-cloud-core==0.25.0
google-cloud-datastore==1.2.0
google-cloud-pubsub==0.26.0
googleapis-common-protos==1.5.3
grpcio==1.12.1
mock==2.0.0
prometheus_client==0.5.0
protobuf==3.5.2.post1
PyYAML==3.12
